requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
langchain-groq>=0.1.4
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
python-dotenv>=1.0.1


#pip install -r requirements.txt
